Almost killed SplineAdoptorBase.
