+++ src/QMCWaveFunctions/BsplineFactory/SplineC2CAdoptor.h
@@ -58,13 +58,7 @@ struct SplineC2CSoA : public SplineAdoptorBase<ST, 3>, public BsplineSet
   using hContainer_type  = VectorSoaContainer<ST, 6>;
   using ghContainer_type = VectorSoaContainer<ST, 10>;
 
-  using Base::first_spo;
   using Base::GGt;
-  using Base::kPoints;
-  using Base::last_spo;
-  using Base::MakeTwoCopies;
-  using Base::offset;
-  using Base::PrimLattice;
 
   ///multi bspline set
   std::shared_ptr<MultiBspline<ST>> SplineInst;
@@ -83,16 +77,16 @@ struct SplineC2CSoA : public SplineAdoptorBase<ST, 3>, public BsplineSet
 
   SplineC2CSoA() : Base()
   {
-    this->is_complex   = true;
-    this->AdoptorName  = "SplineC2CSoAAdoptor";
-    this->KeyWord      = "SplineC2CSoA";
+    is_complex  = true;
+    AdoptorName = "SplineC2CSoAAdoptor";
+    KeyWord     = "SplineC2CSoA";
   }
 
   SPOSet* makeClone() const override { return new SplineC2CSoA(*this); }
 
   inline void resizeStorage(size_t n, size_t nvals)
   {
-    Base::init_base(n);
+    init_base(n);
     size_t npad = getAlignedSize<ST>(2 * n);
     myV.resize(npad);
     myG.resize(npad);
@@ -150,7 +144,7 @@ struct SplineC2CSoA : public SplineAdoptorBase<ST, 3>, public BsplineSet
   bool read_splines(hdf_archive& h5f)
   {
     std::ostringstream o;
-    o << "spline_" << SplineAdoptorBase<ST, D>::MyIndex;
+    o << "spline_" << MyIndex;
     einspline_engine<SplineType> bigtable(SplineInst->getSplinePtr());
     return h5f.readEntry(bigtable, o.str().c_str()); //"spline_0");
   }
@@ -158,7 +152,7 @@ struct SplineC2CSoA : public SplineAdoptorBase<ST, 3>, public BsplineSet
   bool write_splines(hdf_archive& h5f)
   {
     std::ostringstream o;
-    o << "spline_" << SplineAdoptorBase<ST, D>::MyIndex;
+    o << "spline_" << MyIndex;
     einspline_engine<SplineType> bigtable(SplineInst->getSplinePtr());
     return h5f.writeEntry(bigtable, o.str().c_str()); //"spline_0");
   }
